+++ src/Command/Redis/SHUTDOWN.php
@@ -26,4 +26,36 @@ class SHUTDOWN extends RedisCommand
     {
         return 'SHUTDOWN';
     }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setArguments(array $arguments)
+    {
+        if (empty($arguments)) {
+            parent::setArguments($arguments);
+
+            return;
+        }
+
+        $processedArguments = [];
+
+        if (array_key_exists(0, $arguments) && null !== $arguments[0]) {
+            $processedArguments[] = ($arguments[0]) ? 'SAVE' : 'NOSAVE';
+        }
+
+        if (array_key_exists(1, $arguments) && false !== $arguments[1]) {
+            $processedArguments[] = 'NOW';
+        }
+
+        if (array_key_exists(2, $arguments) && false !== $arguments[2]) {
+            $processedArguments[] = 'FORCE';
+        }
+
+        if (array_key_exists(3, $arguments) && false !== $arguments[3]) {
+            $processedArguments[] = 'ABORT';
+        }
+
+        parent::setArguments($processedArguments);
+    }
 }