+++ src/Connection/Factory.php
@@ -13,6 +13,7 @@
 namespace Predis\Connection;
 
 use InvalidArgumentException;
+use Predis\Client;
 use Predis\Command\RawCommand;
 use ReflectionClass;
 use UnexpectedValueException;
@@ -171,6 +172,16 @@ class Factory implements FactoryInterface
 
             $connection->addConnectCommand(
                 new RawCommand('AUTH', $cmdAuthArgs)
+            );
+        }
+
+        if ($parameters->client_info ?? false && !$connection instanceof RelayConnection) {
+            $connection->addConnectCommand(
+                new RawCommand('CLIENT', ['SETINFO', 'LIB-NAME', 'predis'])
+            );
+
+            $connection->addConnectCommand(
+                new RawCommand('CLIENT', ['SETINFO', 'LIB-VER', Client::VERSION])
             );
         }