+++ tests/Predis/ClientTest.php
@@ -1253,6 +1253,36 @@ class ClientTest extends PredisTestCase
         $this->assertSame('127.0.0.1:6381', $iterator->key());
     }
 
+    /**
+     * @group connected
+     * @group relay-incompatible
+     * @requiresRedisVersion >= 7.2.0
+     */
+    public function testDoNoSetClientInfoOnConnection(): void
+    {
+        $client = new Client($this->getParameters());
+        $libName = $client->client('LIST')[0]['lib-name'];
+        $libVer = $client->client('LIST')[0]['lib-ver'];
+
+        $this->assertEmpty($libName);
+        $this->assertEmpty($libVer);
+    }
+
+    /**
+     * @group connected
+     * @group relay-incompatible
+     * @requiresRedisVersion >= 7.2.0
+     */
+    public function testSetClientInfoOnConnectionWhenEnabled(): void
+    {
+        $client = new Client($this->getParameters(['client_info' => true]));
+        $libName = $client->client('LIST')[0]['lib-name'];
+        $libVer = $client->client('LIST')[0]['lib-ver'];
+
+        $this->assertSame('predis', $libName);
+        $this->assertSame(Client::VERSION, $libVer);
+    }
+
     // ******************************************************************** //
     // ---- HELPER METHODS ------------------------------------------------ //
     // ******************************************************************** //