+++ tests/Predis/Command/Redis/BloomFilter/BFLOADCHUNK_Test.php
@@ -61,6 +61,7 @@ class BFLOADCHUNK_Test extends PredisCommandTestCase
 
     /**
      * @group connected
+     * @group relay-resp3
      * @return void
      * @requiresRedisBfVersion >= 1.0.0
      */