+++ tests/Predis/Command/Redis/HGETALL_Test.php
@@ -77,16 +77,40 @@ class HGETALL_Test extends PredisCommandTestCase
 
     /**
      * @group connected
+     * @group cluster
+     * @requiresRedisVersion >= 6.0.0
+     */
+    public function testReturnsAllTheFieldsAndTheirValuesUsingCluster(): void
+    {
+        $redis = $this->getClient();
+
+        $redis->del('metavars');
+
+        $this->testReturnsAllTheFieldsAndTheirValues();
+    }
+
+    /**
+     * @group connected
      * @requiresRedisVersion >= 2.0.0
      */
     public function testThrowsExceptionOnWrongType(): void
     {
         $this->expectException('Predis\Response\ServerException');
-        $this->expectExceptionMessage('Operation against a key holding the wrong kind of value');
+        $this->expectExceptionMessageMatches('/.*Operation against a key holding the wrong kind of value.*/');
 
         $redis = $this->getClient();
 
         $redis->set('foo', 'bar');
         $redis->hgetall('foo');
+    }
+
+    /**
+     * @group connected
+     * @group cluster
+     * @requiresRedisVersion >= 6.0.0
+     */
+    public function testThrowsExceptionOnWrongTypeUsingCluster(): void
+    {
+        $this->testThrowsExceptionOnWrongType();
     }
 }