+++ tests/Predis/Command/Redis/Search/FTALIASADD_Test.php
@@ -63,6 +63,7 @@ class FTALIASADD_Test extends PredisCommandTestCase
 
     /**
      * @group connected
+     * @group relay-resp3
      * @return void
      * @requiresRediSearchVersion >= 1.0.0
      */
@@ -100,6 +101,7 @@ class FTALIASADD_Test extends PredisCommandTestCase
 
     /**
      * @group connected
+     * @group relay-resp3
      * @return void
      * @requiresRediSearchVersion >= 1.0.0
      */