+++ tests/Predis/Command/Redis/Search/FTCREATE_Test.php
@@ -62,6 +62,7 @@ class FTCREATE_Test extends PredisCommandTestCase
 
     /**
      * @group connected
+     * @group relay-resp3
      * @return void
      * @requiresRediSearchVersion >= 1.0.0
      */