+++ tests/Predis/Command/Redis/Search/FTDICTDEL_Test.php
@@ -60,6 +60,7 @@ class FTDICTDEL_Test extends PredisCommandTestCase
 
     /**
      * @group connected
+     * @group relay-resp3
      * @dataProvider dictionariesProvider
      * @param  array $addArguments
      * @param  array $deleteArguments