+++ tests/Predis/Command/Redis/Search/FTSPELLCHECK_Test.php
@@ -62,6 +62,7 @@ class FTSPELLCHECK_Test extends PredisCommandTestCase
 
     /**
      * @group connected
+     * @group relay-resp3
      * @return void
      * @requiresRediSearchVersion >= 1.4.0
      */