+++ tests/Predis/Command/Redis/TDigest/TDIGESTBYRANK_Test.php
@@ -61,6 +61,7 @@ class TDIGESTBYRANK_Test extends PredisCommandTestCase
 
     /**
      * @group connected
+     * @group relay-resp3
      * @return void
      * @requiresRedisBfVersion >= 2.4.0
      */